feat(customer): centralize validation schemas in shared package
