+++ packages/shared/src/schemas/index.ts
@@ -1,2 +1,3 @@
-export * from './roles.schema';
 export * from './users.schema';
+export * from './roles.schema';
+export * from './customers.schema';
